requirements.txt
repoze.sendmail==4.1
sqlalchemy>=1.0.12
ziggurat_foundations>=0.6.7
celery>=3.1.18
redis>=2.10.3
psycopg2>=2.6
paginate>=0.5.4
paginate_sqlalchemy
pyramid>=1.7b3
pyramid_jinja2
pyramid_debugtoolbar
pyramid_tm
pyramid_mailer
pyramid_redis_sessions>=1.0.1
pyramid_authstack
transaction
zope.sqlalchemy
waitress
dogpile.cache
marshmallow
wtforms
alembic
authomatic
six
structlog
cryptography
webhelpers2_grid
appenlight_client
requests
